Removed JPA Added a static MockDatabase.class in the App.class which will function as the ''Database''.
Implemented the Beans functionalities. Next step is to add the Rest endpoints and test the functionalities. After this, create the front end.

In the Beans the MockDatabase is used. Due to the ID's not being generated automatically, a working solution has been added:
The ID of a new item is the size of the list in which it is placed, +1.

This is a working solution, however programming wise a very ugly solution. In JPA this would be automated, due to the database not working this is the fastest solution.