- Completed TweetBean.java - Completed UserBean.java - Created and executed tests for all TweetBean methods with Postman - Created and executed tests for all UserBean methods with Postman - Added correct Rest annotations in the App.java class (credit to Rowan for noticing this missing piece of code) - Changed map structure (added unique map path, credits to Edwin van Rooij for suggesting) - Changed MockDatabase.java usage (credits to Rowan for suggesting) The MockDatabase is now called through a wrapper method in MockingBean.java (MockingBean.getDb()) instead through App.java (App.getDb())
All functionalities work except:
- Logging in:
Somehow the String comparison does not pass, unclear as the values are exactly the same even in ASCII values
- Time of Tweets
They do work except they don't show the Amsterdam time (UTC+02:00). This has to be looked into as in, will we change this in back-end or front-end code (what is possible?).